cleaning up and starting nmc/block endpt
